requirements.txt
# HyperLiquid Perp Trading Bot Requirements
# =======================================
# Updated: December 2025

# Core Python packages
asyncio-mqtt>=0.16.0
aiohttp>=3.11.0
aiofiles>=24.1.0
websockets>=14.0

# HyperLiquid official SDK
hyperliquid-python-sdk>=0.21.0

# Data processing and analysis
pandas>=2.2.0
numpy>=2.2.0
scipy>=1.14.0

# Technical analysis
ta-lib>=0.6.0
pandas-ta>=0.3.14b0

# Configuration management
pyyaml>=6.0.2
python-dotenv>=1.1.0
pydantic>=2.12.0

# Logging and monitoring
loguru>=0.7.3
structlog>=25.1.0

# Database (PostgreSQL for analytics and AI dataset)
sqlalchemy>=2.0.36
asyncpg>=0.31.0
psycopg2-binary>=2.9.10

# Machine Learning (for future V2+)
scikit-learn>=1.6.0
xgboost>=3.1.0
torch>=2.5.0

# Telegram notifications
python-telegram-bot>=21.10

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0

# Development tools
black>=24.10.0
flake8>=7.1.0

# Performance
uvloop>=0.21.0

# Cryptography and security
cryptography>=44.0.0

# HTTP requests
requests>=2.32.0

# Time and date utilities
python-dateutil>=2.9.0

# Environment variables
python-decouple>=3.8